fix: handle detail
